requirements.txt
json
logging
pickle
python-dotenv
redis
telegram
telethon
